Move derive_parent_mass_from_precursor_mz to separate function
